+++ src/Core/Renderer.cpp
@@ -162,7 +162,7 @@ void Renderer::render() {
         int cell_scale = (int)((cam_size * cam_size * cam_size) * 30 + 1);
         shader_base->uniform_mat("_view", 4, &view[0][0]);
         shader_base->uniform_mat("_proj", 4, &proj[0][0]);
-        shader_base->uniform_f("_cansize", scn->info_.width, scn->info_.height);
+        shader_base->uniform_f("_cansize", (float)scn->info_.width, (float)scn->info_.height);
         // shader_base->uniform_f("_size", (float)scn->info_.width, (float)scn->info_.height);
         shader_base->uniform_i("_scale", cell_scale);
         batch_.draw_batch();
@@ -170,7 +170,7 @@ void Renderer::render() {
     
     auto shader_canvas = app_->RES->GetShader("canvas");
     shader_canvas->bind();
-    shader_canvas->uniform_f("_cansize", scn->info_.width, scn->info_.height);
+    shader_canvas->uniform_f("_cansize", (float)scn->info_.width, (float)scn->info_.height);
     shader_canvas->uniform_mat("_view", 4, &view[0][0]);
     shader_canvas->uniform_mat("_proj", 4, &proj[0][0]);
     if (current_paint_tex_ != nullptr) {// draw canvas
